Navbar server-loading updates
